Fix build with CFLAGS -Wextra warning reports
* build no video (PJ_HAS_VIDEO=0)

+++ pjlib/src/pjlib-test/ioq_udp.c
@@ -1243,7 +1243,7 @@ int udp_ioqueue_test()
     pj_bool_t concurs[] = { PJ_TRUE, PJ_FALSE };
     int i, rc, err = 0;
 
-    for (i=0; i<PJ_ARRAY_SIZE(epoll_flags); ++i) {
+    for (i=0; i<(int)PJ_ARRAY_SIZE(epoll_flags); ++i) {
         pj_ioqueue_cfg cfg;
 
         pj_ioqueue_cfg_default(&cfg);
@@ -1257,7 +1257,7 @@ int udp_ioqueue_test()
             err = rc;
     }
 
-    for (i=0; i<PJ_ARRAY_SIZE(concurs); ++i) {
+    for (i=0; i<(int)PJ_ARRAY_SIZE(concurs); ++i) {
         pj_ioqueue_cfg cfg;
 
         pj_ioqueue_cfg_default(&cfg);
@@ -1272,7 +1272,7 @@ int udp_ioqueue_test()
     }
 
 #if PJ_HAS_THREADS
-    for (i=0; i<PJ_ARRAY_SIZE(epoll_flags); ++i) {
+    for (i=0; i<(int)PJ_ARRAY_SIZE(epoll_flags); ++i) {
         pj_ioqueue_cfg cfg;
 
         pj_ioqueue_cfg_default(&cfg);
